Refactoring `self_quality_` and renaming
- Moved `in_left_arg` and `in_right_arg` from `univalence` to `quality`
- Added `sesh_in_left_arg` and `sesh_in_right_arg`
- Shortened `self_quality_left` to `left`
- Shortened `self_quality_right` to `right`
